+++ app/routes/blog/rss[.]xml.tsx
@@ -1,35 +1,10 @@
 import type { LoaderFunction } from '@remix-run/node'
-import * as dateFns from 'date-fns'
+import { formatDate, getHost } from '~/utils/misc'
 
 import { getAllBlogPosts } from '~/utils/post.server'
 
 function cdata(str: string) {
   return `<![CDATA[${str}]]>`
-}
-
-function getHost(request: Request) {
-  const host =
-    request.headers.get('X-Forwarded-Host') ?? request.headers.get('host')
-
-  if (!host) {
-    throw new Error('Could not determine domain URL.')
-  }
-
-  const protocol = host.includes('localhost') ? 'http' : 'https'
-  return `${protocol}://${host}`
-}
-
-function parseDate(dateString: string) {
-  return dateFns.add(dateFns.parseISO(dateString), {
-    minutes: new Date().getTimezoneOffset(),
-  })
-}
-
-function formatDate(dateString: string | Date, format = 'PPP') {
-  if (typeof dateString !== 'string') {
-    dateString = dateString.toISOString()
-  }
-  return dateFns.format(parseDate(dateString), format)
 }
 
 export const loader: LoaderFunction = async ({ request }) => {
